Create EntityHelper class and remove shared folder
