+++ src/directives.ts
@@ -7,12 +7,12 @@
  * return { reply: 'View' }
  */
 
-import { dialog, Response } from "ask-sdk-model";
+import { dialog, events, Response } from "ask-sdk-model";
 import * as bluebird from "bluebird";
+import { promises } from "dns";
 import * as _ from "lodash";
 import { DialogflowReply } from ".";
 import { AlexaReply } from "../src/platforms/alexa/AlexaReply";
-import { generateEntity } from "../src/platforms/shared/directives";
 import { VoxaPlatform } from "./platforms/VoxaPlatform";
 import { ITransition } from "./StateMachine";
 import { IVoxaEvent } from "./VoxaEvent";
@@ -31,6 +31,12 @@ export interface IDirective {
     event: IVoxaEvent,
     transition: ITransition,
   ) => Promise<void>;
+}
+
+export enum EntityOverrideMode {
+  Unspecified = "ENTITY_OVERRIDE_MODE_UNSPECIFIED",
+  Override = "ENTITY_OVERRIDE_MODE_OVERRIDE",
+  Supplement = "ENTITY_OVERRIDE_MODE_SUPPLEMENT",
 }
 
 export function sampleOrItem(
@@ -202,13 +208,162 @@ export class TextP implements IDirective {
   }
 }
 
-export class Entity implements IDirective {
+export abstract class EntityHelper {
+  public generateEntity(entity: any, event: IVoxaEvent): any {
+    const source = _.get(event, "rawEvent.originalDetectIntentRequest.source");
+    const platformName = _.get(event, "platform.name");
+    const platform = source || platformName;
+
+    let newSessionEntity;
+
+    newSessionEntity = entity.reduce(
+      (filteredEntity: any, property: any): any => {
+        let newEntity;
+        let entityMode = _.get(property, "updateBehavior", "REPLACE");
+
+        const name = _.get(property, "name");
+        const entities = _.get(property, "entities");
+
+        this.validateEntityName(name);
+        this.validateEntity(entities);
+
+        if (platform === "google") {
+          entityMode = _.get(
+            property,
+            "entityOverrideMode",
+            "ENTITY_OVERRIDE_MODE_OVERRIDE",
+          );
+
+          this.validateEntityBehavior(entityMode, platform);
+
+          newEntity = this.dialogflowSessionEntity(
+            property,
+            entityMode,
+            name,
+            event,
+          );
+        }
+
+        if (platform === "alexa") {
+          newEntity = this.alexaDynamicEntity(property, name);
+        }
+
+        filteredEntity.push(newEntity);
+        return filteredEntity;
+      },
+      [],
+    );
+
+    if (platform === "alexa") {
+      const behavior =
+        _.chain(entity)
+          .map((e) => e.updateBehavior)
+          .find()
+          .value() || "REPLACE";
+
+      this.validateEntityBehavior(behavior, platform);
+      return (newSessionEntity = {
+        type: "Dialog.UpdateDynamicEntities",
+        types: newSessionEntity,
+        updateBehavior: behavior,
+      });
+    }
+
+    return newSessionEntity;
+  }
+
+  protected dialogflowSessionEntity(
+    property: any,
+    entityOverrideMode: string,
+    name: string,
+    event: IVoxaEvent,
+  ): any {
+    return {
+      entities: property.entities,
+      entityOverrideMode,
+      name: `${event.rawEvent.session}/entityTypes/${name}`,
+    };
+  }
+
+  protected alexaDynamicEntity(property: any, name: string): any {
+    function entityValues(prop: any) {
+      const entity: any = {};
+      if (_.get(prop, "id")) {
+        entity.id = prop.id;
+      }
+      if (_.get(prop, "synonyms") && _.get(prop, "value")) {
+        entity.name = {
+          synonyms: prop.synonyms,
+          value: prop.value,
+        };
+      }
+      return entity;
+    }
+
+    const values: any = property.entities.map((entity: any) =>
+      entityValues(entity),
+    );
+
+    return {
+      name,
+      values,
+    };
+  }
+
+  protected validateEntityBehavior(behavior: string, platform: string): any {
+    const dialogflowEntityBehaviorList = [
+      EntityOverrideMode.Unspecified,
+      EntityOverrideMode.Override,
+      EntityOverrideMode.Supplement,
+    ];
+
+    const alexaEntityBehaviorList = ["REPLACE", "CLEAR"];
+
+    let behaviorList = alexaEntityBehaviorList;
+    let error =
+      "The updateBehavior is incorrect, please consider use one of the followings: REPLACE or CLEAR";
+
+    if (platform === "google") {
+      behaviorList = dialogflowEntityBehaviorList;
+      error =
+        "The entityOverrideMode is incorrect, please consider use one of the followings: ENTITY_OVERRIDE_MODE_UNSPECIFIED, ENTITY_OVERRIDE_MODE_OVERRIDE or ENTITY_OVERRIDE_MODE_SUPPLEMENT";
+    }
+
+    if (!_.includes(behaviorList, behavior)) {
+      throw new Error(error);
+    }
+  }
+
+  protected validateEntity(entities: any): any {
+    if (!entities || _.isEmpty(entities)) {
+      throw new Error(
+        "The entities property is empty or was not provided, please verify",
+      );
+    }
+  }
+
+  protected validateEntityName(name: string): any {
+    const regexName = new RegExp(/^[A-Z-_]+$/i);
+    if (!name) {
+      throw new Error("A name is required for the Entity");
+    }
+
+    if (!regexName.test(name)) {
+      throw new Error(
+        "The name property for the Entity should be only alphabetic characters, and you can include - or _",
+      );
+    }
+  }
+}
+
+export class Entity extends EntityHelper implements IDirective {
   public static key: string = "entities";
   public static platform: string = "core";
 
   public viewPath?: any | any[];
 
   constructor(viewPath: any | any[]) {
+    super();
     this.viewPath = viewPath;
   }
 
@@ -235,7 +390,7 @@ export class Entity implements IDirective {
       );
     }
 
-    entity = generateEntity(entity, event);
+    entity = this.generateEntity(entity, event);
 
     if (platform === "google") {
       (reply as DialogflowReply).sessionEntityTypes = entity;
