--- src/platforms/shared/index.ts
@@ -1,1 +1,0 @@
-export { Entity } from "./directives";
